Added ability to check if file has already been downloaded.

+++ hypeme.py
@@ -26,6 +26,7 @@ from time import time
 from bs4 import BeautifulSoup
 import json
 import string
+import os
 
 ##############AREA_TO_SCRAPE################
 # This is the general area that you'd 
@@ -139,9 +140,12 @@ class HypeScraper:
         
         download_response = urllib2.urlopen(url)
         filename = "{} - {}.mp3".format(artist, title)
-        mp3_song_file = open(filename, "wb")
-        mp3_song_file.write(download_response.read() )
-        mp3_song_file.close()
+        if os.path.exists(filename):
+          print("File already exists , skipping")
+        else:
+          mp3_song_file = open(filename, "wb")
+          mp3_song_file.write(download_response.read() )
+          mp3_song_file.close()
       except urllib2.HTTPError, e:
             print 'HTTPError = ' + str(e.code) + " trying hypem download url."
       except urllib2.URLError, e:
